+++ app/src/main/java/io/github/wulkanowy/ui/modules/about/AboutFragment.kt
@@ -90,11 +90,6 @@ class AboutFragment : BaseFragment<FragmentAboutBinding>(R.layout.fragment_about
             Triple(getString(R.string.about_privacy), getString(R.string.about_privacy_summary), getCompatDrawable(R.drawable.ic_about_privacy))
         }
 
-    override val systemRes: Triple<String, String, Drawable?>?
-        get() = context?.run {
-            Triple(getString(R.string.about_system), getString(R.string.about_system_summary), getCompatDrawable(R.drawable.ic_more_settings))
-        }
-
     override val titleStringId get() = R.string.about_title
 
     companion object {
@@ -177,13 +172,6 @@ class AboutFragment : BaseFragment<FragmentAboutBinding>(R.layout.fragment_about
 
     override fun openPrivacyPolicy() {
         context?.openInternetBrowser("https://wulkanowy.github.io/polityka-prywatnosci.html", ::showMessage)
-    }
-
-    override fun openSystemSettings() {
-        val intent = Intent(Settings.ACTION_APPLICATION_DETAILS_SETTINGS)
-        val uri: Uri = Uri.fromParts("package", requireActivity().packageName, null)
-        intent.data = uri
-        requireActivity().startActivity(intent)
     }
 
     override fun onDestroyView() {
