+++ app/src/main/java/io/github/wulkanowy/ui/modules/about/AboutView.kt
@@ -23,8 +23,6 @@ interface AboutView : BaseView {
 
     val privacyRes: Triple<String, String, Drawable?>?
 
-    val systemRes: Triple<String, String, Drawable?>?
-
     fun initView()
 
     fun updateData(data: List<Triple<String, String, Drawable?>>)
@@ -48,6 +46,4 @@ interface AboutView : BaseView {
     fun openCreators()
 
     fun openPrivacyPolicy()
-
-    fun openSystemSettings()
 }
