+++ app/src/main/java/io/github/wulkanowy/ui/modules/settings/notifications/NotificationsFragment.kt
@@ -1,10 +1,16 @@
 package io.github.wulkanowy.ui.modules.settings.notifications
 
+import android.annotation.SuppressLint
+import android.content.Intent
 import android.content.SharedPreferences
+import android.net.Uri
+import android.os.Build
 import android.os.Bundle
+import android.provider.Settings
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.annotation.RequiresApi
 import androidx.appcompat.app.AlertDialog
 import androidx.preference.Preference
 import androidx.preference.PreferenceFragmentCompat
@@ -16,6 +22,7 @@ import io.github.wulkanowy.R
 import io.github.wulkanowy.ui.base.BaseActivity
 import io.github.wulkanowy.ui.base.ErrorDialog
 import io.github.wulkanowy.ui.modules.main.MainView
+import io.github.wulkanowy.utils.AppInfo
 import io.github.wulkanowy.utils.openInternetBrowser
 import javax.inject.Inject
 
@@ -26,6 +33,9 @@ class NotificationsFragment : PreferenceFragmentCompat(),
 
     @Inject
     lateinit var presenter: NotificationsPresenter
+
+    @Inject
+    lateinit var appInfo: AppInfo
 
     companion object {
         fun newInstance() = NotificationsFragment()
@@ -43,6 +53,13 @@ class NotificationsFragment : PreferenceFragmentCompat(),
 
             setOnPreferenceClickListener {
                 presenter.onFixSyncIssuesClicked()
+                true
+            }
+        }
+
+        findPreference<Preference>(getString(R.string.pref_key_notifications_system_settings))?.run {
+            setOnPreferenceClickListener {
+                presenter.onOpenSystemSettingsClicked()
                 true
             }
         }
@@ -116,6 +133,20 @@ class NotificationsFragment : PreferenceFragmentCompat(),
                 }
             }
             .show()
+    }
+
+    @SuppressLint("InlinedApi")
+    override fun openSystemSettings() {
+        val intent: Intent
+        if (appInfo.systemVersion >= Build.VERSION_CODES.O) {
+            intent = Intent(Settings.ACTION_APP_NOTIFICATION_SETTINGS)
+            intent.putExtra("android.provider.extra.APP_PACKAGE", requireActivity().packageName);
+        } else {
+            intent = Intent(Settings.ACTION_APPLICATION_DETAILS_SETTINGS)
+            val uri: Uri = Uri.fromParts("package", requireActivity().packageName, null)
+            intent.data = uri
+        }
+        requireActivity().startActivity(intent)
     }
 
     override fun onResume() {
